requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
matplotlib==3.8.2
pandas==2.1.4
plotly==5.18.0
kaleido==0.2.1
pytz==2023.3.post1
python-dotenv